feat(ui): introduce mobile-first navigation, dashboard, and voting updates (#70)
# Motivation

We have introduced a new navigation layout that prioritizes a
mobile-first approach. Additionally, we have migrated the content of
certain pages, including account information.


https://github.com/user-attachments/assets/906556b5-253a-4dad-bb8d-d2be625cae0a


https://github.com/user-attachments/assets/8e526f42-05f0-4a0f-9179-d5309365a53a

Note: There is a bug with the Vaul component with long content. I added
a temporal solution
[here](https://github.com/dfinity/governance-app/pull/70/files#diff-0967fffeef1fd0c87afb7405d4528dfdad5a87a291e2040af075d6f73c877b7eR55)
but to be further investigated later.

# Changes

- A new navigation layout has been added, featuring tabs for mobile and
a sidebar for desktop.
-  A Dashboard page with the Account card has been introduced.
-  Proposals have been added to the Voting page.
-  Settings and theme selection options have been included.