feat(dev): simplify local replica setup with system-canisters (#79)
# Motivation

The latest stable version of
[dfx](https://github.com/dfinity/sdk/releases) supports the
`system-canisters` flag, which enables a much simpler local replica.

This PR updates the local instructions to utilize this new feature. A
follow-up PR will implement this approach in the CI, rendering the old
approach obsolete and eligible for garbage collection.

# Changes

- Introduced new scripts to download and start the pocketIc instance
with `system-canisters`.
- Added a `config.json` file to track the version of IC, ensuring that
local and CI environments remain in sync.
- Updated `DEVELOP.md` with the new instructions.